# codec: pre-alloc encode buffer in row level