--- a/util/codec/codec.go
+++ b/util/codec/codec.go
@@ -42,7 +42,7 @@ const (
 	maxFlag          byte = 250
 )
 
-func preRealloc(b []byte, vals []types.Datum, comparable bool, hash bool) []byte {
+func preRealloc(b []byte, vals []types.Datum, comparable bool) []byte {
 	var size int
 	for i := range vals {
 		switch vals[i].Kind() {
@@ -68,7 +68,7 @@ func preRealloc(b []byte, vals []types.Datum, comparable bool, hash bool) []byte
 // encode will encode a datum and append it to a byte slice. If comparable is true, the encoded bytes can be sorted as it's original order.
 // If hash is true, the encoded bytes can be checked equal as it's original value.
 func encode(sc *stmtctx.StatementContext, b []byte, vals []types.Datum, comparable bool, hash bool) (_ []byte, err error) {
-	b = preRealloc(b, vals, comparable, hash)
+	b = preRealloc(b, vals, comparable)
 	for i, length := 0, len(vals); i < length; i++ {
 		switch vals[i].Kind() {
 		case types.KindInt64: