<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/tidb/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->

the first part to fix  #11202

### What is changed and how it works?

pre-calculate buffer and realloc once

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

 - Unit test
 - Integration test

Code changes

 - N/A

Side effects

 - N/A

Related changes

 - N/A

Benchmark

benchmark with 178 columns table + replace into test(this test doesn't use prepare and plan cache if so will got better result)

before

```
[ 5s ] thds: 32 tps: 569.83 qps: 1146.04 (r/w/o: 576.21/569.83/0.00) lat (ms,95%): 106.75 err/s: 0.00 reconn/s: 0.00
[ 10s ] thds: 32 tps: 596.25 qps: 1192.50 (r/w/o: 596.25/596.25/0.00) lat (ms,95%): 95.81 err/s: 0.00 reconn/s: 0.00
[ 15s ] thds: 32 tps: 610.01 qps: 1219.42 (r/w/o: 609.41/610.01/0.00) lat (ms,95%): 97.55 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 32 tps: 599.60 qps: 1198.40 (r/w/o: 598.80/599.60/0.00) lat (ms,95%): 101.13 err/s: 0.00 reconn/s: 0.00
[ 25s ] thds: 32 tps: 588.16 qps: 1177.73 (r/w/o: 589.56/588.16/0.00) lat (ms,95%): 102.97 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 32 tps: 613.64 qps: 1226.89 (r/w/o: 613.24/613.64/0.00) lat (ms,95%): 99.33 err/s: 0.00 reconn/s: 0.00
[ 35s ] thds: 32 tps: 606.35 qps: 1212.50 (r/w/o: 606.15/606.35/0.00) lat (ms,95%): 106.75 err/s: 0.00 reconn/s: 0.00
```

after

```
[ 5s ] thds: 32 tps: 752.44 qps: 1510.67 (r/w/o: 758.23/752.44/0.00) lat (ms,95%): 90.78 err/s: 0.00 reconn/s: 0.00
[ 10s ] thds: 32 tps: 712.86 qps: 1425.32 (r/w/o: 712.46/712.86/0.00) lat (ms,95%): 82.96 err/s: 0.00 reconn/s: 0.00
[ 15s ] thds: 32 tps: 750.60 qps: 1501.21 (r/w/o: 750.60/750.60/0.00) lat (ms,95%): 87.56 err/s: 0.00 reconn/s: 0.00
[ 20s ] thds: 32 tps: 728.80 qps: 1458.00 (r/w/o: 729.20/728.80/0.00) lat (ms,95%): 90.78 err/s: 0.00 reconn/s: 0.00
[ 25s ] thds: 32 tps: 707.99 qps: 1416.39 (r/w/o: 708.39/707.99/0.00) lat (ms,95%): 89.16 err/s: 0.00 reconn/s: 0.00
[ 30s ] thds: 32 tps: 746.20 qps: 1492.60 (r/w/o: 746.40/746.20/0.00) lat (ms,95%): 84.47 err/s: 0.00 reconn/s: 0.00
[ 35s ] thds: 32 tps: 754.41 qps: 1508.61 (r/w/o: 754.21/754.41/0.00) lat (ms,95%): 86.00 err/s: 0.00 reconn/s: 0.00
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/pingcap/tidb/11218)
<!-- Reviewable:end -->