@lysu 2019-07-11
/run-all-tests